Add `path` template literal tag function
This can be used to correctly encode the path for a request. (The usage
example should probably be changed to involve a request function, rather
than console.log(), once request functions exist.)

I’m not yet completely sure if the escaping is correct, because upon
trying to test it I quickly discovered a bug in the REST API where it
doesn’t correctly redirect. Once T411738 has been fixed (and the fix has
been deployed in Wikimedia production), I should try out what the
actually expected URL for "?! (album)" is, and potentially adjust the
test case.