Check HTTP response status for server error (5xx)
Introduce a new Error class for server errors; I don’t think we need
specific classes beyond that. The body is documented as being either a
string or an object; currently it’ll always turn out to be an object
(m3api forcibly tries to decode it, and if it’s not JSON, it’ll throw a
SyntaxError – I intend to fix this in the future internal interface
redesign) but in future we can hopefully preserve non-JSON bodies (most
likely bare HTML) here.