Add getResponseStatus() function
By taking advantage of the fact that all response values are (now – see
bf7b46aaf8) objects with an identity (and thus garbage-collectable), we
can register the response and its corresponding status in a WeakMap, and
make that available via a function. Once the response is no longer used,
it gets garbage collected and the WeakMap entry vanishes too.

I expect this pattern to also work once HTML-returning request functions
are added; those functions should return some kind of DOM object (not
sure if a Document, a Node, a DocumentFragment or something else) which
can also be used as a WeakMap key.

It gets a bit trickier for plaintext-returning request functions (in
practice, those will be ones returning wikitext, such as the Parsoid
endpoints to transform HTML to wikitext). I think we have at least three
options for the return value of those: they could return an ad-hoc
object like { text: 'wikitext' }, which has an identity; they could
return a boxed String instead of a primitive string; or they could
simply not support returning the status at all, on the assumption that
no plaintext-returning endpoint has successful statuses beside 200 OK.
I’m not yet sure which option I’ll go for (this is some time away
anyway, as it’ll first require a rehaul of the m3api internal interface
for network requests).