Support query params in path`` template tag
I’m also planning to add a separate parameter for query params to
getJson(), but this will still be useful for other request methods that
won’t have such a parameter (because the body parameters will be
separate, and unlike in the Action API they’re not interchangeable). See
also my comments on T411325, especially T411325#11438540.