Interpolate params into path
This adds another way to specify path params: put them next to the other
params. If there are any unsubstituted params left in the path (and I
don’t think it’s possible for {} sequences to be anything else – if they
were a “regular” part of the URL they’d be encoded), we pull them out of
the params (throwing an error if they’re not present) and make the
actual request with a copy of the params sans the path params. (We don’t
modify the original params object, just in case someone needs it to stay
stable.)

One use case for this is an extension of the TypeScript pattern
mentioned in the commit message of 011926a393: now in principle all
calls can have a constant string literal as the path parameter (I
think), making it possible to match overloads with the right param
types.

For the syntax (a template expression is "{", then one or more
characters except "{" or "}", then "}"), see section 4.8.2 Path
Templating of the OpenAPI Specification v3.2.0 [1]. If I read the spec
correctly, it’s theoretically possible to have parameters with the same
name in the path and the body (“A unique parameter is defined by a
combination of a name and location”), but I don’t expect that to happen
in practice in MediaWiki or extensions, and in any case it’s not an
issue with this approach: if you have the same param in the path and the
query/body, you just can’t use this feature and have to use path``
tagged templates after all (and then there’s no path param left by the
time the request function is reached).

[1]: https://spec.openapis.org/oas/v3.2.0.html#path-templating