Write README.md documentation
In places, this gets awkward, because I expect certain statements to
hold true for future functions that don’t actually exist yet. For the
most part, I’ve opted to phrase the text in ways that will still be
valid in the future, even if it makes them awkward at the moment (such
as “e.g. `getJson`” when it’s in fact the only case and not just one
example), to reduce the risk that I’ll forget to update the text later.